Fix absolute paths with protocol (i.e. msc://)

+++ src/megatron/energon/epathlib/epath.py
@@ -126,6 +126,14 @@ class EPath:
         with self.open("rb") as f:
             return f.read()
 
+    def write_text(self, text: str) -> None:
+        with self.open("w") as f:
+            f.write(text)
+
+    def write_bytes(self, data: bytes) -> None:
+        with self.open("wb") as f:
+            f.write(data)
+
     def copy(self, target: "EPath") -> None:
         """Copy a file to a new path, possibly between different file systems.
 
@@ -133,17 +141,21 @@ class EPath:
             target: The path to the local file to download to.
         """
 
-        if self.fs == target.fs:
-            self.fs.copy(self._internal_str_path, target._internal_str_path)
-        elif target.is_local():
-            # Note, src and target are swapped for these calls.
-            self.fs.download_file(target._internal_str_path, self._internal_str_path)
-        elif self.is_local():
-            # Note, src and target are swapped for these calls.
-            self.fs.upload_file(target._internal_str_path, self._internal_str_path)
+        if self.is_file():
+            if self.fs == target.fs:
+                self.fs.copy(self._internal_str_path, target._internal_str_path)
+            elif target.is_local():
+                self.fs.download_file(self._internal_str_path, target._internal_str_path)
+            elif self.is_local():
+                target.fs.upload_file(target._internal_str_path, self._internal_str_path)
+            else:
+                with self.open("rb") as src_f, target.open("wb") as dst_f:
+                    shutil.copyfileobj(src_f, dst_f)
         else:
-            with self.open("rb") as src_f, target.open("wb") as dst_f:
-                shutil.copyfileobj(src_f, dst_f)
+            inner_path = EPath(self)
+            for fpath in self.fs.list(self._internal_str_path):
+                inner_path.internal_path = PurePosixPath("/" + fpath.key)
+                inner_path.copy(target / inner_path.relative_to(self))
 
     @property
     def name(self) -> str:
@@ -209,7 +221,7 @@ class EPath:
         if isinstance(other, EPath):
             # Always absolute
             return other
-        if other.startswith("/"):
+        if other.startswith("/") or "://" in other:
             return EPath(other)
 
         new_path = EPath(self)
